moving hacker entry to Evaluator
